requirements.txt
Django==2.0
djangorestframework==3.9.2
django-filter==2.1.0
dj-database-url
psycopg2
gunicorn
coreapi
